mkLine componet, global css @MkList, UI improve UI
